Decouple restli request uri creation so it's pluggable by consumers
